narrow search test scope to just _include
